+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 2024-02-21 - 0.0.11
+
+Fixes:
+
+- Reap zombie processes and respawn on failure (fixes [#8](https://github.com/exacheck/exacheck/issues/8))
+
+Misc:
+
+- Update Sentry and dnspython releases
+
 ## 2024-02-08 - 0.0.10
 
 Changes: